add runtime checks to lesspipe.sh
git-svn-id: svn://svn.code.sf.net/p/lesspipe/code/trunk@3 08e7c0e0-42c1-4d7e-ade3-037309d14f20

+++ lesspipe.sh
@@ -35,6 +35,11 @@ tarcmd=gtar
 if [[ `tar --version 2>&1` = *GNU* ]]; then
   tarcmd=tar
 fi
+
+cmd_exist () {
+  command -v "$1" > /dev/null 2>&1 && return 0 || return 1
+}
+
 filecmd='file -L -s';
 sep=:						# file name separator
 altsep==					# alternate separator character
@@ -43,12 +48,23 @@ if [[ -f "$1" && "$1" = *$sep* || "$1" = *$altsep ]]; then
   xxx="${1%=}"
   set "$xxx"
 fi
-tmpdir=$(mktemp -d -p "${TMPDIR:-/tmp}" lesspipe.XXXXXXXXXX)
+if cmd_exist mktemp; then
+  tmpdir=$(mktemp -d "${TMPDIR:-/tmp}/lesspipe.XXXXXXXXXX")
 
-nexttmp () {
-  # nexttmp -d returns a directory
-  mktemp -p "$tmpdir" $1
-}
+  nexttmp () {
+    # nexttmp -d returns a directory
+    mktemp $1 "${tmpdir}/XXXXXXXX"
+  }
+else
+  tmpdir=${TMPDIR:-/tmp}/lesspipe.$RANDOM
+  mkdir $tmpdir
+
+  nexttmp () {
+    new="$tmpdir/lesspipe.$RANDOM"
+    [[ "$1" = -d ]] && mkdir $new
+    echo $new
+  }
+fi
 [[ -d "$tmpdir" ]] || exit 1
 trap "rm -rf '$tmpdir'" 0
 trap - PIPE
@@ -172,10 +188,10 @@ show () {
 get_cmd () {
   cmd=
   typeset t
-  if [[ "$1" = *[blg]zip*compress* || "$1" = *compress\'d\ * || "$1" = *packed\ data* || "$1" = *LZMA*compress* ]]; then ## added '#..then' to fix vim's syntax parsing
+  if [[ "$1" = *[blg]zip*compress* || "$1" = *compress\'d\ * || "$1" = *packed\ data* || "$1" = *LZMA*compress* || "$1" = *xz*compress* ]]; then ## added '#..then' to fix vim's syntax parsing
     if [[ "$3" = $sep$sep ]]; then
       return
-    elif [[ "$1" = *bzip*compress* ]]; then
+    elif [[ "$1" = *bzip*compress* ]] && cmd_exist bzip2; then
       cmd=(bzip2 -cd "$2")
       if [[ "$2" != - ]]; then filen="$2"; fi
       case "$filen" in
@@ -183,20 +199,6 @@ get_cmd () {
         *.tbz) filen="${filen%.tbz}.tar";;
       esac
       return
-    elif [[ "$1" = *lzip*compress* ]]; then
-      cmd=(lzip -cd "$2")
-      if [[ "$2" != - ]]; then filen="$2"; fi
-      case "$filen" in
-        *.lz) filen="${filen%.lz}";;
-        *.tlz) filen="${filen%.tlz}.tar";;
-      esac
-    elif [[ "$1" = *LZMA*compress* ]]; then
-      cmd=(lzma -cd "$2")
-      if [[ "$2" != - ]]; then filen="$2"; fi
-      case "$filen" in
-        *.lzma) filen="${filen%.lzma}";;
-        *.tlz) filen="${filen%.tlz}.tar";;
-      esac
     elif [[ "$1" = *gzip\ compress* || "$1" =  *compress\'d\ * || "$1" = *packed\ data* ]]; then ## added '#..then' to fix vim's syntax parsing
       cmd=(gzip -cd "$2")
       if [[ "$2" != - ]]; then filen="$2"; fi
@@ -222,21 +224,17 @@ get_cmd () {
         istemp "ar p" "$2" data.tar.gz | gzip -dc - > "$t"
       fi
       cmd=(istar "$t" "$file2")
-    elif [[ "$1" = *RPM* ]]; then
+    elif [[ "$1" = *RPM* ]] && cmd_exist cpio && ( cmd_exist rpm2cpio || cmd_exist rpmunpack ); then
       cmd=(isrpm "$2" "$file2")
-    elif [[ "$1" = *Jar\ archive* ]]; then
+    elif [[ "$1" = *Jar\ archive* ]] && cmd_exist fastjar; then
       cmd=(isjar "$2" "$file2")
-    elif [[ "$1" = *Zip* || "$1" = *ZIP* ]]; then
+    elif [[ "$1" = *Zip* || "$1" = *ZIP* ]] && cmd_exist unzip; then
       cmd=(istemp "unzip -avp" "$2" "$file2")
-    elif [[ "$1" = *RAR\ archive* ]]; then
-      cmd=(istemp "unrar p -inul" "$2" "$file2")
-    elif [[ "$1" = *7-zip\ archive* ]]; then
-      cmd=(istemp "7za e -so" "$2" "$file2")
-    elif [[ "$1" = *[Cc]abinet* ]]; then
+    elif [[ "$1" = *[Cc]abinet* ]] && cmd_exist cabextract; then
       cmd=(iscab "$2" "$file2")
     elif [[ "$1" = *\ ar\ archive* ]]; then
       cmd=(istemp "ar p" "$2" "$file2")
-    elif [[ "$1" = *ISO\ 9660* ]]; then
+    elif [[ "$1" = *ISO\ 9660* ]] && cmd_exist isoinfo; then
       cmd=(isoinfo "-i$2" "-x$file2")
     fi
     if [[ "$cmd" != "" ]]; then
@@ -259,15 +257,6 @@ istar () {
   $tarcmd Oxf "$1" "$2" 2>/dev/null
 }
 
-isdvi () {
-  typeset t
-  if [[ "$1" != *.dvi ]]; then
-    t="$tmpdir/tmp.dvi"
-    cat "$1" > "$t"
-    set "$t"
-  fi
-  dvi2tty -q "$1"
-}
 
 istemp () {
   typeset prog
@@ -304,18 +293,28 @@ nodash () {
 }
 
 isrpm () {
-  typeset t
-  if [[ "$1" = - ]]; then
-    t=$(nexttmp)
-    cat > "$t"
-    set "$t" "$2"
+  if cmd_exist rpm2cpio && cmd_exist cpio; then
+    typeset t
+    if [[ "$1" = - ]]; then
+      t=$(nexttmp)
+      cat > "$t"
+      set "$t" "$2"
+    fi
+    # setup $b as a batch file containing "$b.out"
+    typeset b
+    b=$(nexttmp -d)
+    echo "$b.out" > "$b"
+    # to support older versions of cpio the --to-stdout option is not used here
+    #rpm2cpio "$1"|cpio -i --quiet --rename-batch-file "$b" "$2"
+    #cat "$b.out"
+    rpm2cpio "$1"|pax -rw "$2" $b
+    cat "$b/$2"
+  elif cmd_exist rpmunpack && cmd_exist cpio; then
+    # rpmunpack will write to stdout if it gets file from stdin
+    # extract file $2 from archive $1, assume that cpio is sufficiently new
+    # (option --to-stdout existing) if rpmunpack is installed
+    cat "$1" | rpmunpack | gzip -cd | cpio -i --quiet --to-stdout "$2"
   fi
-  # setup $b as a batch file containing "$b.out"
-  typeset b
-  b=$(nexttmp)
-  echo "$b.out" > "$b"
-  rpm2cpio "$1"|cpio -i --quiet --rename-batch-file "$b" "$2"
-  cat "$b.out"
 }
 
 isjar () {
@@ -339,9 +338,26 @@ isjar () {
   )
 }
 
-PARSEHTML=yes
-parsehtml () { nodash "elinks -dump -force-html" "$1"; }
+if cmd_exist html2text || cmd_exist elinks || cmd_exist links || cmd_exist lynx || cmd_exist w3m; then
+  PARSEHTML=yes
+else
+  PARSEHTML=no
+fi
 #parsexml () { nodash "elinks -dump -default-mime-type text/xml" "$1"; }
+parsehtml () {
+  if [[ "$PARSEHTML" = no ]]; then
+    echo "==> No suitable tool for HTML parsing found, install one of html2text, elinks, links, lynx or w3m"
+    return
+  elif cmd_exist elinks; then
+    nodash "elinks -dump -force-html" "$1"
+  elif cmd_exist links; then
+    if [[ "$1" = - ]]; then set - -stdin; fi
+    links -dump -force_html "$1"
+  elif cmd_exist lynx; then
+    if [[ "$1" = - ]]; then set - -stdin; fi
+    lynx -dump -force_html "$1" && return
+  fi
+}
 
 isfinal() {
   typeset t
@@ -367,11 +383,24 @@ isfinal() {
     echo "==> use tar_file${sep}contained_file to view a file in the archive"
     $tarcmd tvf "$2"
   elif [[ "$1" = *RPM* ]]; then
-    echo "==> use RPM_file${sep}contained_file to view a file in the RPM"
-    istemp "rpm -qivp" "$2"
-    echo "================================= Content ======================================"
-    istemp rpm2cpio "$2"|cpio -i -tv --quiet
-  elif [[ "$1" = *roff* ]]; then
+    header="==> use RPM_file${sep}contained_file to view a file in the RPM"
+    if cmd_exist rpm; then
+      echo $header
+      istemp "rpm -qivp" "$2"
+      header="";
+    fi
+    if cmd_exist cpio && cmd_exist rpm2cpio; then
+      echo $header
+      echo "================================= Content ======================================"
+      istemp rpm2cpio "$2"|cpio -i -tv 2>/dev/null
+    elif cmd_exist cpio && cmd_exist rpmunpack; then
+      echo $header
+      echo "================================= Content ======================================"
+      cat "$2" | rpmunpack | gzip -cd | cpio -i -tv 2>/dev/null
+    else
+      echo "==> please install rpm2cpio or rpmunpack to see the contents of RPM files"
+    fi
+  elif [[ "$1" = *roff* ]] && cmd_exist groff; then
     DEV=utf8
     if [[ $LANG != *UTF*8* && $LANG != *utf*8* ]]; then
       if [[ "$LANG" = ja* ]]; then
@@ -396,7 +425,7 @@ isfinal() {
     istemp "ar p" "$2" data.tar.gz | gzip -dc - | $tarcmd tvf -
   # do not display all perl text containing pod using perldoc
   #elif [[ "$1" = *Perl\ POD\ document\ text* || "$1" = *Perl5\ module\ source\ text* ]]; then
-  elif [[ "$1" = *Perl\ POD\ document\ text* ]]; then
+  elif [[ "$1" = *Perl\ POD\ document\ text* ]] && cmd_exist perldoc; then
     echo "==> append $sep to filename to view the perl source"
     istemp perldoc "$2"
   elif [[ "$1" = *\ script* ]]; then
@@ -404,76 +433,62 @@ isfinal() {
   elif [[ "$1" = *text\ executable* ]]; then
     set "plain text" "$2"
   elif [[ "$1" = *PostScript* ]]; then
-    echo "==> append $sep to filename to view the postscript file"
-    istemp ps2ascii "$2"
+    if cmd_exist pstotext; then
+      echo "==> append $sep to filename to view the postscript file"
+      nodash pstotext "$2"
+    elif cmd_exist ps2ascii; then
+      echo "==> append $sep to filename to view the postscript file"
+      istemp ps2ascii "$2"
+    else
+      echo "==> install pstotext or ps2ascii to view a textual representation of the file contents"
+    fi
   elif [[ "$1" = *executable* ]]; then
     echo "==> append $sep to filename to view the binary file"
     nodash strings "$2"
   elif [[ "$1" = *\ ar\ archive* ]]; then
     echo "==> use library${sep}contained_file to view a file in the archive"
     istemp "ar vt" "$2"
-  elif [[ "$1" = *shared* ]]; then
+  elif [[ "$1" = *shared* ]] && cmd_exist nm; then
     echo "==> This is a dynamic library, showing the output of nm"
     istemp nm "$2"
-  elif [[ "$1" = *Jar\ archive* ]]; then
+  elif [[ "$1" = *Jar\ archive* ]] && cmd_exist fastjar; then
     echo "==> use jar_file${sep}contained_file to view a file in the archive"
     nodash "fastjar -tf" "$2"
-  elif [[ "$1" = *Zip* || "$1" = *ZIP* ]]; then
+  elif [[ "$1" = *Zip* || "$1" = *ZIP* ]] && cmd_exist unzip; then
     echo "==> use zip_file${sep}contained_file to view a file in the archive"
     istemp "unzip -lv" "$2"
-  elif [[ "$1" = *RAR\ archive* ]]; then
-    echo "==> use rar_file${sep}contained_file to view a file in the archive"
-    istemp "unrar v" "$2"
-  elif [[ "$1" = *7-zip\ archive* ]]; then
-    typeset res
-    res=$(istemp "7za l" "$2")
-    if [[ "$res" = *\ 1\ file* ]]; then
-      echo "==> a 7za archive containing one file was silently unpacked"
-      if [[ "$2" != - ]]; then
-        7za e -so "$2" 2>/dev/null
-      else
-        # extract name of temporary file containing the 7za archive
-        t=${res#*Listing\ archive:\ }
-        t2="
-"
-        t=${t%%$t2*}
-        7za e -so $t 2>/dev/null
-      fi
-    else
-      echo "==> use 7za_file${sep}contained_file to view a file in the archive"
-      echo "$res"
-    fi
-  elif [[ "$1" = *[Cc]abinet* ]]; then
+  elif [[ "$1" = *[Cc]abinet* ]] && cmd_exist cabextract; then
     echo "==> use cab_file${sep}contained_file to view a file in the cabinet"
     istemp "cabextract -l" "$2"
-  elif [[ "$1" = *\ DVI* ]]; then
-    echo "==> append $sep to filename to view the binary DVI file"
-    isdvi "$2"
   elif [[ "$PARSEHTML" = yes && "$1" = *HTML* ]]; then
     echo "==> append $sep to filename to view the HTML source"
     parsehtml "$2"
-  elif [[ "$PARSEHTML" = yes && "$1" = *PDF* ]]; then
+  elif [[ "$PARSEHTML" = yes && "$1" = *PDF* ]] && cmd_exist pdftohtml; then
     echo "==> append $sep to filename to view the PDF source"
     t=$(nexttmp)
     cat "$2" > "$t"; pdftohtml -stdout "$t" | parsehtml -
-  elif [[ "$1" = *PDF* ]]; then
+  elif [[ "$1" = *PDF* ]] && cmd_exist pdftotext; then
     echo "==> append $sep to filename to view the PDF source"
     istemp pdftotext "$2" -
   elif [[ "$1" = *Microsoft\ Word* || "$1" = *Microsoft\ Office* ]]; then
-    echo "==> append $sep to filename to view the raw word document"
-    antiword "$2"
-  elif [[ "$1" = *Rich\ Text\ Format* ]]; then
-    if [[ "$PARSEHTML" = yes ]]; then
-      echo "==> append $sep to filename to view the RTF source"
-      istemp "unrtf --html" "$2" | parsehtml -
+    if cmd_exist antiword; then
+      echo "==> append $sep to filename to view the raw word document"
+      antiword "$2"
+    elif cmd_exist catdoc; then
+      echo "==> append $sep to filename to view the raw word document"
+      catdoc "$2"
     else
-      echo "==> append $sep to filename to view the RTF source"
-      istemp "unrtf --text" "$2" | sed -e "s/^### .*//" | fmt -s
+      echo "==> install antiword or catdoc to view human readable text"
+      cat "$2"
     fi
-  elif [[ "$1" = *OpenDocument\ [CHMPST]* || "$1" = *OpenOffice\.org\ 1\.x\ [CIWdgpst]* ]]; then
-    echo "==> append $sep to filename to view the OpenOffice or OpenDocument source"
-    istemp sxw2txt "$2"
-  elif [[ "$1" = *ISO\ 9660* ]]; then
+  elif [[ "$PARSEHTML" = yes && ("$1" = *OpenDocument\ [CHMPST]* || "$1" = *OpenOffice\.org\ 1\.x\ [CIWdgpst]*) ]] && cmd_exist unzip; then
+    if cmd_exist sxw2txt; then
+      echo "==> append $sep to filename to view the OpenOffice or OpenDocument source"
+      istemp sxw2txt "$2"
+    else
+      echo "==> install at least sxw2txt from the lesspipe package to see plain text in openoffice documents"
+    fi
+  elif [[ "$1" = *ISO\ 9660* ]] && cmd_exist isoinfo; then
     if [[ "$2" != - ]]; then
       echo "==> append $sep to filename to view the binary data"
       isoinfo -d -i "$2"
@@ -481,28 +496,25 @@ isfinal() {
       echo "================================= Content ======================================"
       isoinfo -lR$joliet -i "$2"
     fi
-  elif [[ "$1" = *image\ data*  || "$1" = *image\ text* || "$1" = *JPEG\ file* || "$1" = *JPG\ file* ]]; then
+  elif [[ "$1" = *image\ data*  || "$1" = *image\ text* || "$1" = *JPEG\ file* || "$1" = *JPG\ file* ]] && cmd_exist identify; then
     echo "==> append $sep to filename to view the binary data"
     identify -verbose "$2"
-  elif [[ "$1" = *MPEG\ *layer\ 3\ audio* || "$1" = *MPEG\ *layer\ III* || "$1" = *mp3\ file* || "$1" = *MP3* ]]; then
-    echo "==> append $sep to filename to view the binary data"
-    istemp "id3v2 -l" "$2"
   elif [[ "$1" = *perl\ Storable* ]]; then
     echo "==> append $sep to filename to view the binary data"
     perl -MStorable=retrieve -MData::Dumper -e '$Data::Dumper::Indent=1;print Dumper retrieve shift' "$2"
-  elif [[ "$1" = *UTF-8* && $LANG != *UTF*8* && $LANG != *utf*8* ]]; then
+  elif [[ "$1" = *UTF-8* && $LANG != *UTF*8* && $LANG != *utf*8* ]] && cmd_exist iconv -c; then
     echo "==> append $sep to filename to view the UTF-8 encoded data"
     iconv -c -f UTF-8 -t ISO-8859-1 "$2"
-  elif [[ "$1" = *ISO-8859* && ($LANG = *UTF*8* || $LANG = *utf*8*) ]]; then
+  elif [[ "$1" = *ISO-8859* && ($LANG = *UTF*8* || $LANG = *utf*8*) ]] && cmd_exist iconv -c; then
     echo "==> append $sep to filename to view the ISO-8859 encoded data"
     iconv -c -f ISO-8859-1 -t UTF-8 "$2"
-  elif [[ "$1" = *UTF-16* && $LANG != *UTF*8* && $LANG != *utf*8* ]]; then
+  elif [[ "$1" = *UTF-16* && $LANG != *UTF*8* && $LANG != *utf*8* ]] && cmd_exist iconv -c; then
     echo "==> append $sep to filename to view the UTF-16 encoded data"
     iconv -c -f UTF-16 -t ISO-8859-1 "$2"
-  elif [[ "$1" = *UTF-16* && ($LANG = *UTF*8* || $LANG = *utf*8*) ]]; then
+  elif [[ "$1" = *UTF-16* && ($LANG = *UTF*8* || $LANG = *utf*8*) ]] && cmd_exist iconv -c; then
     echo "==> append $sep to filename to view the UTF-16 encoded data"
     iconv -c -f UTF-16 -t UTF-8 "$2"
-  elif [[ "$1" = *GPG\ encrypted\ data* ]]; then
+  elif [[ "$1" = *GPG\ encrypted\ data* ]] && cmd_exist gpg; then
     echo "==> append $sep to filename to view the encrypted file"
     gpg -d "$2"
   elif [[ "$1" = *data* ]]; then
@@ -519,10 +531,13 @@ isfinal() {
 IFS=$sep a="$@"
 IFS=' '
 if [[ "$a" = "" ]]; then
+  if [[ "$0" != /* ]]; then
+     pat=`pwd`/
+  fi
   if [[ "$SHELL" = *csh ]]; then
-    echo "setenv LESSOPEN \"|$0 %s\""
+    echo "setenv LESSOPEN \"|$pat$0 %s\""
   else
-    echo "LESSOPEN=\"|$0 %s\""
+    echo "LESSOPEN=\"|$pat$0 %s\""
     echo "export LESSOPEN"
   fi
 else
